+++ docx4j-core-tests/src/test/java/org/docx4j/fonts/FontsTestSupport.java
@@ -80,6 +80,11 @@ final class FontsTestSupport {
 			ThemePart tp = new ThemePart();
 			tp.setJaxbElement((org.docx4j.dml.Theme) XmlUtils.unwrap(XmlUtils.unmarshalString(themeXml(fontSchemeXml))));
 			pkg.getMainDocumentPart().addTargetPart(tp);
+		} else {
+			// since 17.1.1 createPackage adds the Office theme part, so "no theme part"
+			// has to be made rather than assumed
+			pkg.getMainDocumentPart().getRelationshipsPart()
+					.removePart(new org.docx4j.openpackaging.parts.PartName("/word/theme/theme1.xml"));
 		}
 		pkg.getMainDocumentPart().setJaxbElement((org.docx4j.wml.Document) XmlUtils.unwrap(XmlUtils.unmarshalString(
 				"<w:document xmlns:w=\"" + W + "\"><w:body>" + bodyXml + "</w:body></w:document>")));